Refactor PaymentMethod Configuration
