+++ Sources/PrimerSDK/Classes/User Interface/TokenizationViewModels/CardFormPaymentMethodTokenizationViewModel.swift
@@ -352,7 +352,7 @@ class CardFormPaymentMethodTokenizationViewModel: PaymentMethodTokenizationViewM
         }
     }
     
-    required init(config: PaymentMethodConfig) {
+    required init(config: PaymentMethod.Configuration) {
         self.flow = Primer.shared.flow.internalSessionFlow.vaulted ? .vault : .checkout
         super.init(config: config)
